requirements.txt
runpod
loguru
pillow
boto3
